Résolution de bug de compatibilité et ajout de commentaires et de documentation

See merge request louis.bahrman/groupe3-tictactoe!26

+++ src/tictactoecodingame/ArbreMinMax.java
@@ -9,18 +9,23 @@ import java.util.ArrayList;
 
 /**
  *
- * @author senda
+ * <p>Structure d'arbre pour l'algoritme MinMax.</p><p>Cette structure est récursive et peut être un noeud ou un élément vide.
+ * Les noeuds référencent un ou plusieurs coups, une valeur évaluant la qualité dues coups et une liste pointant vers d'autrs noeuds qui sont 
+ * ses fils.</p>
  */
 public class ArbreMinMax {
     
+    //La valeur est stockée dans un double pour povoir est modulée facilement par l'amélioration de mémorisation
     protected double value;
     //Coup théorique joué à ce noeud
     protected Coup coup;
+    //Si on fait plusieurs coups
     protected ArrayList<Coup> coups;
     protected ArrayList<ArbreMinMax> fils;
     
     // Les constructeurs :
     
+    //L'abre vide est représenté par des attributs vides
     public ArbreMinMax(){   
     }
     
@@ -75,7 +80,7 @@ public class ArbreMinMax {
     }
     
     
-    //Des choses sans nom :
+    //Les méthodes de réglage des attributs :
     public void setvalue(double value){
         this.value = value;
     }
@@ -113,8 +118,14 @@ public class ArbreMinMax {
         }
     }
     
+    //fonctions auxiliaires pour l'algo MinMax
+    
+    /**
+     * <div>Fonction déterminant la valeur minimum parmi les noeuds fils</div>
+     * @return 
+     */
     public double Min(){
-        double m = Double.MAX_VALUE;
+        double m = (double)Integer.MAX_VALUE;
         int a = this.getfils().size();
         for(int i = 0; i < a ; i++){
             double n = this.getfils().get(i).getvalue();
@@ -125,8 +136,12 @@ public class ArbreMinMax {
         return m;
     }
     
+    /**
+     * <div>Fonction déterminant la valeur maximum parmi les noeuds fils</div>
+     * @return 
+     */
     public double Max(){
-        double m = Double.MIN_VALUE;
+        double m = (double)Integer.MIN_VALUE;
         int a = this.getfils().size();
         for(int i = 0; i < a ; i++){
             double n = this.getfils().get(i).getvalue();
@@ -137,6 +152,12 @@ public class ArbreMinMax {
         return m;
     }
     
+    /**
+     * <div>Algorithme du Minmax. L'algo fonctione récursivement, on mémorise la profondeur à laquelle on se situe
+     * pour savoir si on remonte la valeur maximale où minimale. A la racine on veut toukours le maximum puisqu'on évalue du point
+     * de vue du joeuru cible. Ensuite la parité de la profondeur permet déterminer le joueur qui s'apprêtre à jouer</div>
+     * @param c Porofndeur en cours dans le parcours de l'arbre
+     */
     public void MinMax(int c){
         // c = compteur 
         // Le compteur doit être initialisé à 0 donc pair -> Max, impair -> Min
